+++ pywigner/__init__.py
@@ -1,0 +1,9 @@
+import operators
+
+def git_rev_actual():
+    from subprocess import check_output
+    import os.path
+    git_dir = os.path.dirname(os.path.realpath(__file__))
+    return check_output(["git", "-C", git_dir, "rev-parse", "HEAD"])[:-1]
+    # chops the newline at the end
+
