+++ pywigner/operators/__init__.py
@@ -1,0 +1,2 @@
+from operators import Operator
+from coherent_states import CoherentProjection
