+++ pywigner/tools.py
@@ -1,0 +1,7 @@
+def clean_ravel(arr, n_dofs):
+    try:
+        retval = arr.ravel()
+    except AttributeError:
+        retval = [arr] * n_dofs
+    return retval
+
